Completed Message

+++ src/main/java/Service/MessageService.java
@@ -5,6 +5,8 @@ import Model.Message;
 
 import java.util.List;
 
+import javax.management.RuntimeErrorException;
+
 public class MessageService {
     MessageDAO messageDAO;
 
@@ -19,6 +21,12 @@ public class MessageService {
         if(message.getMessage_text().length() > 255){
             throw new IllegalArgumentException("Message cannot be more than 255 characters");
         }
+        if (message.getPosted_by() <= 0) {
+            throw new IllegalArgumentException("Invalid user ID");
+        }
+        if (message.getTime_posted_epoch() <= 0) {
+            throw new IllegalArgumentException("Invalid timestamp");
+        }
         return messageDAO.insertMessage(message);
     }
 
@@ -27,21 +35,37 @@ public class MessageService {
     }
 
     public Message getMessageById(int id){
-        return messageDAO.getMessageById(id);
+        Message message = messageDAO.getMessageById(id);
+        if(message == null){
+            throw new IllegalArgumentException("Message not found with the id");
+        }
+        return message;
     }
 
     public void deleteMessage(int id){
+        Message message = getMessageById(id);
+        if(message == null){
+            throw new IllegalArgumentException("Message with ID " + id + " does not exist and cannot be deleted.");
+        }
         messageDAO.deleteMessage(id);
     }
 
     public Message updateMessage(int id, Message message){
+        if(message.getMessage_text() == null || message.getMessage_text().isBlank()){
+            throw new IllegalArgumentException("Message cannot be null or blank");
+        }
+        if(message.getMessage_text().length() > 255){
+            throw new IllegalArgumentException("Message cannot be more than 255 characters");
+        }
+        if (message.getPosted_by() <= 0) {
+            throw new IllegalArgumentException("Invalid user ID");
+        }
         try {
             messageDAO.updateMessage(id, message);
             Message updatedMessage = messageDAO.getMessageById(id);
             return updatedMessage; 
         } catch (Exception e) {
-            System.out.print("Error while updating the account: " + e.getMessage());
-            return null;
+            throw new RuntimeException("Error while updating the account: " + e.getMessage());
         }
     }
